feat: Trades service redis done

+++ trade_service/model/entity/trades.entity.go
@@ -8,7 +8,7 @@ import (
 
 type Trades struct {
 	ID        int            `json:"id" gorm:"primaryKey"`
-	Price     int            `json:"price" gorm:"index"`
+	Price     float64        `json:"price" gorm:"index"`
 	CoinID    int            `json:"coin_id" gorm:"index"`
 	UserID    int            `json:"user_id" gorm:"index"`
 	CreatedAt time.Time      `json:"created_at"`
