+++ trade_service/model/request/trades.request.go
@@ -1,6 +1,6 @@
 package request
 
 type TradeCreateRequest struct {
-	Price  int `json:"price" validate:"required"`
-	CoinID int `json:"coin_id" validate:"required"`
+	Price  float64 `json:"price" validate:"required"`
+	CoinID int     `json:"coin_id" validate:"required"`
 }
